hw02/manuscript_paper-main/script_ntut/requirements.txt
numpy==1.21.0 
matplotlib==3.4.2
svglib==1.5.1
reportlab==3.6.12
pikepdf==8.5.2
Pillow==10.1.0
tqdm==4.64.1
